+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - \#297 NPE on Startup
 - \#298 Configure Action for Double Click Build
 - Expand Job Node after Load Build for first time
+- \#301 ExtensibleChoiceParameterDefinition support request
+- \#159 "Unknown parameter" for "Extensible Choice" parameters
 
 ## [0.13.11]
 - Rename Plugin to 'Jenkins Control' to respect Jetbrains naming convention